Add support for project level (current dir) .zprintrc
ISSUE 3 This will team zprint-clj how to find a .zprintrc file in the
current directory.  The idea is to allow users to have zprint-clj find
and respect project level configurations.  This is a naive solution.  A
better solution would be to have zprint-clj be smarter about how it find
the .zprintrc file